# docs/openapi/serviceability-api.yaml
openapi: 3.0.3
info:
  title: Prayog Serviceability API
  description: |
    Comprehensive API for managing service area coverage, geographical entities, location management, and delivery serviceability.

    ## Features
    - **Geographical Management**: Countries, regions, districts, cities, and areas
    - **Location Types & Aliases**: Flexible location classification and aliasing system
    - **Postal Code Management**: Postal code management with location scope validation
    - **Partner Integration**: External partner service integration and validation
    - **Serviceability Check**: Service area coverage and delivery options

    ## Authentication
    All endpoints require valid authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer <your-api-key>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    {
      "status": "success",
      "data": { ... }
    }
    ```

    Error responses follow this format:
    ```json
    {
      "status": "error", 
      "message": "Error description",
      "code": "error_code"
    }
    ```
  version: 1.0.0
  contact:
    email: api@prayog.com
    name: Prayog API Support
  license:
    name: Proprietary
    url: https://prayog.com/license
servers:
  - url: "https://api.prayog.com"
    description: Production server
  - url: "https://api.staging.prayog.com"
    description: Staging server
  - url: "http://localhost:8080"
...
security:
  - BearerAuth: []
paths:
...
  /serviceability/v1/check:
    post:
      tags:
        - serviceability
      summary: Check serviceability for a location
      description: Check if a postal code or origin-destination pair is serviceable
      operationId: checkServiceability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceabilityCheckRequest"
      responses:
        "200":
          description: Serviceability check completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceabilityResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/bulk-check:
    post:
      tags:
        - serviceability
      summary: Bulk serviceability check
      description: Check serviceability for multiple locations in a single request
      operationId: bulkCheckServiceability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkServiceabilityRequest"
      responses:
        "200":
          description: Bulk serviceability check completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkServiceabilityResponse"
        "207":
          description: Multi-status - partial success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkServiceabilityResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /serviceability/v1/partners/{partner_id}/location-coverages:
    parameters:
      - name: partner_id
        in: path
        required: true
        description: Partner UUID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - partner-location-coverage
      summary: List partner location coverages
      description: Retrieve all location coverages for a specific partner with optional filtering
      operationId: listPartnerLocationCoverages
      parameters:
        - name: location_scope
          in: query
          description: Filter by location scope
          schema:
            type: string
            enum: [POSTAL_CODE, AREA, CITY, REGION, COUNTRY]
        - name: location_id
          in: query
          description: Filter by location ID
          schema:
            type: string
            format: uuid
        - name: zone_type
          in: query
          description: Filter by zone type
          schema:
            type: string
            enum: [PRIMARY, SECONDARY, BUFFER]
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Successfully retrieved partner location coverages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerLocationCoverageListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - partner-location-coverage
      summary: Create partner location coverage
      description: Create a new location coverage area for a partner
      operationId: createPartnerLocationCoverage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerLocationCoverageRequest"
      responses:
        "201":
          description: Partner location coverage created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerLocationCoverageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/partners/{partner_id}/location-coverages/bulk:
    parameters:
      - name: partner_id
        in: path
        required: true
        description: Partner UUID
        schema:
          type: string
          format: uuid
    post:
      tags:
        - partner-location-coverage
      summary: Bulk create partner location coverages
      description: Create multiple location coverage areas for a partner in a single request
      operationId: bulkCreatePartnerLocationCoverages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreatePartnerLocationCoverageRequest"
      responses:
        "201":
          description: Bulk creation completed (check individual results for success/failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCreatePartnerLocationCoverageResponse"
        "207":
          description: Multi-status - partial success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCreatePartnerLocationCoverageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /serviceability/v1/partners/{partner_id}/location-coverages/{coverage_id}:
    parameters:
      - name: partner_id
        in: path
        required: true
        description: Partner UUID
        schema:
          type: string
          format: uuid
      - name: coverage_id
        in: path
        required: true
        description: Location coverage UUID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - partner-location-coverage
      summary: Get partner location coverage
      description: Retrieve a specific location coverage for a partner
      operationId: getPartnerLocationCoverage
      responses:
        "200":
          description: Successfully retrieved partner location coverage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerLocationCoverageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - partner-location-coverage
      summary: Update partner location coverage
      description: Update an existing location coverage for a partner
      operationId: updatePartnerLocationCoverage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerLocationCoverageRequest"
      responses:
        "200":
          description: Partner location coverage updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerLocationCoverageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - partner-location-coverage
      summary: Delete partner location coverage
      description: Delete a location coverage area for a partner
      operationId: deletePartnerLocationCoverage
      responses:
        "204":
          description: Partner location coverage deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # Region Types
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  parameters:
...
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        code:
          type: string
    PaginationInfo:
      type: object
      properties:
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0

    # Country Schemas
...
    ServiceabilityCheckRequest:
      type: object
      description: Request for checking serviceability of a location or route
      properties:
        postal_code:
          type: string
          description: Postal code for single location check
          example: "110001"
        pickup_postal_code:
          type: string
          description: Pickup postal code for origin-destination check
          example: "110001"
        delivery_postal_code:
          type: string
          description: Delivery postal code for origin-destination check
          example: "560001"
        country_code:
          type: string
          pattern: "^[A-Z]{2}$"
          description: ISO 3166-1 alpha-2 country code
          example: "IN"
          default: "IN"
        service_types:
          type: array
          items:
            type: string
            enum: [STANDARD, EXPRESS, OVERNIGHT, SAME_DAY]
          description: List of service types to check
          example: ["STANDARD", "EXPRESS"]
        partner_ids:
          type: array
          items:
            type: string
          description: Optional list of specific partner IDs to check
          example: ["partner-123", "partner-456"]
    BulkServiceabilityRequest:
      type: object
      description: Request for bulk serviceability checking
      required:
        - requests
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceabilityCheckRequest"
          minItems: 1
          maxItems: 100
          description: Array of serviceability check requests
    ServiceabilityResponse:
      type: object
      description: Response for serviceability check
      required:
        - success
      properties:
        success:
          type: boolean
          description: Whether the request was processed successfully
        data:
          $ref: "#/components/schemas/ServiceabilityData"
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    BulkServiceabilityResponse:
      type: object
      description: Response for bulk serviceability check
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: Whether all requests were processed successfully
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceabilityResponse"
          description: Array of individual serviceability responses
    ServiceabilityData:
      type: object
      description: Serviceability data for a location or route
      properties:
        serviceable:
          type: boolean
          description: Whether the location/route is serviceable
        services:
          type: array
          items:
            type: object
            properties:
              service_type:
                type: string
              partner_id:
                type: string
              service_available:
                type: boolean
              estimated_delivery_days:
                type: integer
          description: Available services for the location/route

    # Partner Location Coverage Schemas
    PartnerLocationCoverage:
      type: object
      description: Partner location coverage information
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the coverage record
          example: "550e8400-e29b-41d4-a716-446655440000"
        partner_id:
          type: string
          format: uuid
          description: Partner unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        partner_code:
          type: string
          description: Partner code reference
          example: "PARTNER_001"
        location_scope:
          type: string
          enum: [POSTAL_CODE, AREA, CITY, REGION, COUNTRY]
          description: The scope/level of location coverage
          example: "CITY"
        location_id:
          type: string
          format: uuid
          description: Location entity unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        location_code:
          type: string
          description: Location code reference
          example: "NYC"
        zone_type:
          type: string
          enum: [PRIMARY, SECONDARY, BUFFER]
          description: Zone classification for coverage area
          example: "PRIMARY"
        is_active:
          type: boolean
          description: Whether the coverage is currently active
          example: true
        created_at:
          type: string
          format: date-time
          description: Coverage creation timestamp
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Coverage last update timestamp
          example: "2023-01-01T00:00:00Z"
    CreatePartnerLocationCoverageRequest:
      type: object
      description: Request to create a new partner location coverage
      properties:
        location_scope:
          type: string
          enum: [POSTAL_CODE, AREA, CITY, REGION, COUNTRY]
          description: The scope/level of location coverage
          example: "CITY"
        location_id:
          type: string
          format: uuid
          description: Location entity unique identifier (required if location_code not provided)
          example: "550e8400-e29b-41d4-a716-446655440000"
        location_code:
          type: string
          description: Location code reference (required if location_id not provided)
          maxLength: 20
          example: "NYC"
        zone_type:
          type: string
          enum: [PRIMARY, SECONDARY, BUFFER]
          description: Zone classification for coverage area
          example: "PRIMARY"
        is_active:
          type: boolean
          description: Whether the coverage should be active
          default: true
          example: true
    UpdatePartnerLocationCoverageRequest:
      type: object
      description: Request to update an existing partner location coverage
      properties:
        location_scope:
          type: string
          enum: [POSTAL_CODE, AREA, CITY, REGION, COUNTRY]
          description: The scope/level of location coverage
          example: "CITY"
        location_id:
          type: string
          format: uuid
          description: Location entity unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        location_code:
          type: string
          description: Location code reference
          maxLength: 20
          example: "NYC"
        zone_type:
          type: string
          enum: [PRIMARY, SECONDARY, BUFFER]
          description: Zone classification for coverage area
          example: "PRIMARY"
        is_active:
          type: boolean
          description: Whether the coverage should be active
          example: true
    PartnerLocationCoverageResponse:
      type: object
      description: Response containing partner location coverage data
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/PartnerLocationCoverage"
    PartnerLocationCoverageListResponse:
      type: object
      description: Response containing list of partner location coverages
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            coverages:
              type: array
              items:
                $ref: "#/components/schemas/PartnerLocationCoverage"
              description: Array of partner location coverage records
            pagination:
              $ref: "#/components/schemas/PaginationInfo"
    BulkCreatePartnerLocationCoverageRequest:
      type: object
      description: Request to create multiple partner location coverages
      required:
        - coverages
      properties:
        coverages:
          type: array
          items:
            $ref: "#/components/schemas/CreatePartnerLocationCoverageRequest"
          minItems: 1
          maxItems: 100
          description: Array of coverage creation requests
    BulkCreatePartnerLocationCoverageResponse:
      type: object
      description: Response for bulk partner location coverage creation
      properties:
        status:
          type: string
          enum: [success, partial_success, error]
        data:
          type: object
          properties:
            created:
              type: array
              items:
                $ref: "#/components/schemas/PartnerLocationCoverage"
              description: Successfully created coverage records
            failed:
              type: array
              items:
                type: object
                properties:
                  index:
                    type: integer
                    description: Index of the failed request in the input array
                  error:
                    type: string
                    description: Error message for the failed request
                  request:
                    $ref: "#/components/schemas/CreatePartnerLocationCoverageRequest"
              description: Failed coverage creation attempts
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [validation_error]
          example:
            status: error
            message: "Field 'name' is required"
            code: validation_error
    ConflictError:
      description: Resource already exists or constraint violation
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [conflict_error]
          example:
            status: error
            message: "Country with code 'US' already exists"
            code: conflict_error
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [not_found]
          example:
            status: error
            message: "Country not found"
            code: not_found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  code:
                    type: string
                    enum: [internal_error]
          example:
            status: error
            message: "An internal error occurred"
            code: internal_error